Chat bot

+++ app/src/main/java/com/example/sereno/chat/model/ChatModel.kt
@@ -1,6 +1,6 @@
 package com.example.sereno.chat.model
 
-data class ChatModel(val chat: String, val owner: ChatOwner, val dateEpoch: Int)
+data class ChatModel(val chat: String, val owner: ChatOwner, val dateEpoch: Long)
 
 enum class ChatOwner {
     USER, BOT
